requirements.txt
python-telegram-bot[job-queue]==20.6
python-dotenv==1.0.0 